+++ cercle-5/miniRT/srcs/gui/render/optimizations/taa/reproject.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/06 00:00:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/06 00:00:00 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/06 10:16:43 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ static t_vec3	get_wp_taa(t_gui *gui, size_t dx, size_t dy)
 	size_t			pw;
 
 	o = &gui->opts;
-	pw = (size_t)o->prev_render_size.x;
-	ndc = repro_get_ndc(vec2i((int)dx, (int)dy), gui->win.size,
+	pw = o->prev_render_size.x;
+	ndc = repro_get_ndc(vec2i(dx, dy), gui->win.size,
 			vec2(o->cur_half_w, o->cur_half_h));
 	dir = repro_get_dir(o->cur_cam, ndc);
 	return (vec3_add(o->cur_cam.pos,
@@ -42,8 +42,8 @@ bool	reproject_taa(t_gui *gui, size_t dx, size_t dy, t_vec2i *out)
 	r.half = vec2(o->prev_half_w, o->prev_half_h);
 	if (!repro_world_to_screen(r, get_wp_taa(gui, dx, dy), out, &cz))
 		return (false);
-	return (fabs(cz - o->prev_depth[(size_t)out->y * (size_t)r.size.x
-				+ (size_t)out->x]) / fmaxf((float)cz,
-			(float)o->prev_depth[(size_t)out->y * (size_t)r.size.x
-				+ (size_t)out->x]) < TAA_DEPTH_THR);
+	return (fabs(cz - o->prev_depth[out->y * r.size.x
+				+ out->x]) / fmaxf((float)cz,
+			(float)o->prev_depth[out->y * r.size.x
+				+ out->x]) < TAA_DEPTH_THR);
 }
